refactor(ai): colocate catalog service and composable
