+++ src/app/ai/models/catalog/index.ts
